+++ eclipse-collections-api/src/main/java/org/eclipse/collections/api/bag/MutableBag.java
@@ -244,4 +244,7 @@ public interface MutableBag<T>
      */
     @Override
     ImmutableBag<T> toImmutable();
+
+    @Override
+    MutableSet<T> asSet();
 }
