+++ eclipse-collections-api/src/main/java/org/eclipse/collections/api/bag/sorted/ImmutableSortedBag.java
@@ -236,4 +236,7 @@ public interface ImmutableSortedBag<T>
 
     @Override
     ImmutableSortedBag<T> drop(int count);
+
+    @Override
+    ImmutableSortedSet<T> asSet();
 }
