+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/bag/immutable/ImmutableEmptyBag.java
@@ -783,4 +783,16 @@ final class ImmutableEmptyBag<T>
     {
         return Sets.immutable.empty();
     }
+
+    @Override
+    public ImmutableSet<T> asSet()
+    {
+        return Sets.immutable.empty();
+    }
+
+    @Override
+    public RichIterable<T> distinctView()
+    {
+        return this.asSet();
+    }
 }
