+++ unit-tests/src/test/java/org/eclipse/collections/impl/bag/mutable/MultiReaderHashBagAsWriteUntouchableTest.java
@@ -80,4 +80,9 @@ public class MultiReaderHashBagAsWriteUntouchableTest extends AbstractCollection
         MutableSet<String> actual = bag.selectUnique();
         Assert.assertEquals(expected, actual);
     }
+
+    public void asSet()
+    {
+        Verify.assertThrows(UnsupportedOperationException.class, () -> this.newWith().asSet());
+    }
 }
