+++ unit-tests/src/test/java/org/eclipse/collections/impl/set/immutable/ImmutableSetFromBagAdapterTest.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2020 Goldman Sachs and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * and Eclipse Distribution License v. 1.0 which accompany this distribution.
+ * The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
+ * and the Eclipse Distribution License is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ */
+
+package org.eclipse.collections.impl.set.immutable;
+
+import org.eclipse.collections.api.factory.Bags;
+import org.eclipse.collections.api.set.ImmutableSet;
+
+public class ImmutableSetFromBagAdapterTest extends AbstractImmutableSetTestCase
+{
+    @Override
+    protected ImmutableSet<Integer> classUnderTest()
+    {
+        return new ImmutableSetFromBagAdapter<>(Bags.immutable.of(1, 2, 3, 4));
+    }
+}
