+++ unit-tests/src/test/java/org/eclipse/collections/impl/set/sorted/immutable/ImmutableSortedSetFromBagAdapterTest.java
@@ -1,0 +1,128 @@
+/*
+ * Copyright (c) 2020 Goldman Sachs and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * and Eclipse Distribution License v. 1.0 which accompany this distribution.
+ * The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
+ * and the Eclipse Distribution License is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ */
+
+package org.eclipse.collections.impl.set.sorted.immutable;
+
+import java.util.Collections;
+import java.util.Comparator;
+
+import org.eclipse.collections.api.factory.SortedBags;
+import org.eclipse.collections.api.set.sorted.ImmutableSortedSet;
+import org.eclipse.collections.impl.block.factory.PrimitiveFunctions;
+import org.eclipse.collections.impl.list.mutable.primitive.BooleanArrayList;
+import org.eclipse.collections.impl.list.mutable.primitive.ByteArrayList;
+import org.eclipse.collections.impl.list.mutable.primitive.CharArrayList;
+import org.eclipse.collections.impl.list.mutable.primitive.DoubleArrayList;
+import org.eclipse.collections.impl.list.mutable.primitive.FloatArrayList;
+import org.eclipse.collections.impl.list.mutable.primitive.IntArrayList;
+import org.eclipse.collections.impl.list.mutable.primitive.LongArrayList;
+import org.eclipse.collections.impl.list.mutable.primitive.ShortArrayList;
+import org.junit.Assert;
+import org.junit.Test;
+
+public class ImmutableSortedSetFromBagAdapterTest extends AbstractImmutableSortedSetTestCase
+{
+    @Override
+    protected ImmutableSortedSet<Integer> classUnderTest()
+    {
+        return new ImmutableSortedSetFromBagAdapter<>(SortedBags.immutable.of(1, 2, 3, 4));
+    }
+
+    @Override
+    protected ImmutableSortedSet<Integer> classUnderTest(Comparator<? super Integer> comparator)
+    {
+        return new ImmutableSortedSetFromBagAdapter<>(SortedBags.immutable.of(comparator, 1, 2, 3, 4));
+    }
+
+    @Override
+    @Test(expected = UnsupportedOperationException.class)
+    public void subSet()
+    {
+        this.classUnderTest().castToSortedSet().subSet(1, 4);
+    }
+
+    @Override
+    @Test(expected = UnsupportedOperationException.class)
+    public void headSet()
+    {
+        this.classUnderTest().castToSortedSet().headSet(4);
+    }
+
+    @Override
+    @Test(expected = UnsupportedOperationException.class)
+    public void tailSet()
+    {
+        this.classUnderTest().castToSortedSet().tailSet(1);
+    }
+
+    @Override
+    @Test
+    public void collectBoolean()
+    {
+        ImmutableSortedSet<Integer> integers = this.classUnderTest(Collections.reverseOrder());
+        Assert.assertEquals(BooleanArrayList.newListWith(true, true, true, true), integers.collectBoolean(PrimitiveFunctions.integerIsPositive()));
+    }
+
+    @Override
+    @Test
+    public void collectByte()
+    {
+        ImmutableSortedSet<Integer> integers = this.classUnderTest(Collections.reverseOrder());
+        Assert.assertEquals(ByteArrayList.newListWith((byte) 4, (byte) 3, (byte) 2, (byte) 1), integers.collectByte(PrimitiveFunctions.unboxIntegerToByte()));
+    }
+
+    @Override
+    @Test
+    public void collectChar()
+    {
+        ImmutableSortedSet<Integer> integers = this.classUnderTest(Collections.reverseOrder());
+        Assert.assertEquals(CharArrayList.newListWith('D', 'C', 'B', 'A'), integers.collectChar(integer -> (char) (integer.intValue() + 64)));
+    }
+
+    @Override
+    @Test
+    public void collectDouble()
+    {
+        ImmutableSortedSet<Integer> integers = this.classUnderTest(Collections.reverseOrder());
+        Assert.assertEquals(DoubleArrayList.newListWith(4.0d, 3.0d, 2.0d, 1.0d), integers.collectDouble(PrimitiveFunctions.unboxIntegerToDouble()));
+    }
+
+    @Override
+    @Test
+    public void collectFloat()
+    {
+        ImmutableSortedSet<Integer> integers = this.classUnderTest(Collections.reverseOrder());
+        Assert.assertEquals(FloatArrayList.newListWith(4.0f, 3.0f, 2.0f, 1.0f), integers.collectFloat(PrimitiveFunctions.unboxIntegerToFloat()));
+    }
+
+    @Override
+    @Test
+    public void collectInt()
+    {
+        ImmutableSortedSet<Integer> integers = this.classUnderTest(Collections.reverseOrder());
+        Assert.assertEquals(IntArrayList.newListWith(4, 3, 2, 1), integers.collectInt(PrimitiveFunctions.unboxIntegerToInt()));
+    }
+
+    @Override
+    @Test
+    public void collectLong()
+    {
+        ImmutableSortedSet<Integer> integers = this.classUnderTest(Collections.reverseOrder());
+        Assert.assertEquals(LongArrayList.newListWith(4, 3, 2, 1), integers.collectLong(PrimitiveFunctions.unboxIntegerToLong()));
+    }
+
+    @Override
+    @Test
+    public void collectShort()
+    {
+        ImmutableSortedSet<Integer> integers = this.classUnderTest(Collections.reverseOrder());
+        Assert.assertEquals(ShortArrayList.newListWith((short) 4, (short) 3, (short) 2, (short) 1), integers.collectShort(PrimitiveFunctions.unboxIntegerToShort()));
+    }
+}
